requirements.txt
flask

gunicorn

flask_sqlalchemy

psycopg2-binary

python-dotenv